Call issue -
-> the account is used on several clients
-> your android device is in idle mode.
-> someone is calling you
-> your both clients ring
-> you answer to the call from the other client (for the webbrowser for example)

The android device never stopped ringing until the call is finished.
It should be fixed by now : the client is active until the call is finished.

It was triggered because the android client never receives push from its own account so the m.call.answer is not received.
It was the same issue if the user stops / rejects the call from another client.